requirements.txt
Flask==3.1.3
requests==2.32.5
python-dotenv==1.2.1
urllib3==2.6.0
Werkzeug==3.1.6
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
click==8.3.0
blinker==1.9.0
Flask-OIDC==2.4.0
Flask-WTF==1.2.1
authlib==1.6.6
cryptography==46.0.5
tenacity==9.0.0
APScheduler==3.10.4